+++ lib/eager_eye/configuration.rb
@@ -6,7 +6,7 @@ module EagerEye
 
     DEFAULT_DETECTORS = %i[
       loop_association serializer_nesting missing_counter_cache
-      custom_method_query count_in_iteration
+      custom_method_query count_in_iteration callback_query
     ].freeze
 
     def initialize